+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 ## Unreleased
 ### Add
+- Add validation for plugin configuration form
+- Add FactFinder logger
+- Add German translations
 - Add reporting for Shopware marketplace
 - Add PHPStan
 ### Change