+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## Unreleased
+## [v4.4.0] - 2024.03.12
 ### Add
 - Add validation for plugin configuration form
 - Add FactFinder logger
@@ -409,6 +409,7 @@
 ## [v1.0.0] - 2021.06.29
 Initial module release. Includes Web Components v4.0.3
 
+[v4.4.0]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v4.4.0
 [v4.3.2]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v4.3.2
 [v4.3.1]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v4.3.1
 [v4.3.0]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v4.3.0